shell: replace wordToArg with expand package
expand.Fields performs the same logic that a shell does
when expanding a simple command's words into string fields to execute.

Arguably, some of that is unnecessary for this tool;
for example, expand.Field deals with parameter and command substitutions,
which this tool forbids earlier in the code.
However, some of what it does is still relevant, like unquoting
and dealing with escape sequences, which wordToArg only did partially.