shell: use an allowlist to filter allowed syntax
Otherwise it's easy to forget about some like ForClause or PipeAll,
or statements which are to be negated or run in the background.
We still keep a bit of code to make human-friendly errors.

Note that we were making human-friendly errors for some bash nodes,
such as extended globs or quotes with dollar signs,
but we parse in POSIX mode so right now those never appear anyway.

Add tests to cover node types which the previous blocklist missed,
as well as some relatively common bash nodes which we don't parse yet.